chore(logger): Export stream-helpers from logger package
